These changes are for the purpose of configuring
the service principle to only allow login to the application if the user is allowed.
The problem is that with this method of configuration the service principle for the service needs
AAD admin approval.

This might work fine if you have your own AAD tenant, but for large corps, like MS this is
harder to do.

Parking these changes for right now. We can either publish these if we get asked for them or
we can publish a document that describes how to implement it manually via the portal.